### 2025-02-05 #7750 @crvernon
:wave: @wustat - Thank you for your submission to JOSS.  I am going to run this by our larger editorial board to make sure it meets our requirements for substantial scholarly effort.  I'll get back to you ASAP on this.
