+++ manuscript/JOSS/paper.bib
@@ -255,5 +255,6 @@
   title = {{tibble: Simple Data Frames}},
   author = {Kirill Müller and Hadley Wickham},
   year = {2023},
+  doi={10.32614/CRAN.package.tibble},
   url = {https://tibble.tidyverse.org}
 }
