### 2025-02-05 #7750 @wustat
For reviewers, I suggest **nhejazi** and **limengbinggz**.

> **Submitting author:** [@wustat](https://github.com/wustat) ([Wenbo Wu](http://orcid.org/0000-0002-7642-9773)) **Repository:** https://github.com/wustat/rjaf **Branch with paper.md** (empty if default branch): **Version:** v0.1.1 **Editor:** Pending **Reviewers:** Pending **Managing EiC:** Chris Vernon
>
> ## Status
> [![status](https://camo.githubusercontent.com/7c25004fae5d444ab67853ecfb93aa81f56e4700111a28640e6e6abcfe433195/68747470733a2f2f6a6f73732e7468656f6a2e6f72672f7061706572732f66663866613732356363343064303234373135386264323434663131313762652f7374617475732e737667)](https://joss.theoj.org/papers/ff8fa725cc40d0247158bd244f1117be)
>
> Status badge code:
>
> ```
> ```
>
> **Author instructions**
>
> Thanks for submitting your paper to JOSS [@wustat](https://github.com/wustat). **Currently, there isn't a JOSS editor assigned** to your paper.
>
> [@wustat](https://github.com/wustat) if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). You can search [the list of people](https://reviewers.joss.theoj.org/lookup) that have already agreed to review and may be suitable for this submission.
>
> **Editor instructions**
>
> The JOSS submission bot [@editorialbot](https://github.com/editorialbot) is here to help you find and assign reviewers and start the main review. To find out what [@editorialbot](https://github.com/editorialbot) can do for you type:
>
> ```
> @editorialbot commands
> ```
