### 2025-02-12 0413c21
Move renaming trt with ".rjaf" to the end
