### 2025-02-12 56a9134
change return type to a list with res and counterfactuals
