### 2025-02-12 f65fb80
sort unique trt and edit col names for counterfactuals
