+++ r-package/rjaf/man/residualize.Rd
@@ -25,7 +25,7 @@ data for training and estimation with residualized outcomes.
 }
 \description{
 This function employs random forests and cross-validation to residualize
-outcomes following Wu and Gagnon-Bartsch (2018).
+outcomes following Section 3.3 of Ladhania et al. (2023).
 That is, predicted outcomes resulting from random forests are
 subtracted from the original outcomes. Doing so helps in adjusting for small imbalanaces
 in baseline covariates and removing part of the variation in
@@ -42,7 +42,8 @@ Example_resid <- residualize(Example_trainest, y, vars, nfold = 5, fun.rf = "ran
 
 }
 \references{
-Wu, Edward and Johann A Gagnon-Bartsch (2018). The LOOP Estimator: Adjusting
-for Covariates in Randomized Experiments. Evaluation Review, 42(4):458–488.
+Ladhania Rahul, Spiess Jann, Ungar Lyle, Wu Wenbo (2023). Personalized Assignment
+to One of Many Treatment Arms via Regularized and Clustered Joint Assignment Forests.
+https://doi.org/10.48550/arXiv.2311.00577.
 \cr
 }
