### 2025-02-13 2862bf5
update documentation & test-rjaf.R
