+++ r-package/rjaf/man/rjaf.Rd
@@ -119,11 +119,13 @@ If \code{clus.tree.growing} and \code{clus.outcome.avg} are \code{TRUE}, \code{r
 returns a list of two objects: a tibble named as \code{res} consisting of individual
 IDs, cluster identifiers, and predicted outcomes, and a data frame named as
 \code{clustering} consisting of cluster identifiers, probabilities of being assigned
-to the clusters, and treatment arms. Otherwise, \code{rjaf} simply returns a tibble
-of individual IDs (\code{id}), optimal treatment arms identified by the algorithm (\code{trt.rjaf}), treatment
+to the clusters, and treatment arms. Otherwise, \code{rjaf}  returns a list of two tibbles
+named \code{res} and \code{counterfactuals}. \code{res} consists of individual IDs (\code{id}),
+optimal treatment arms identified by the algorithm (\code{trt.rjaf}), treatment
 clusters (\code{clus.rjaf}) if \code{clus.tree.growing} is \code{TRUE}, and predicted optimal outcomes (\code{Y.rjaf}).
 If counterfactual outcomes are also present, they will be included
-in the tibble along with the column of predicted outcomes (\code{Y.cf}).
+in \code{res} along with the column of predicted outcomes (\code{Y.cf}). \code{counterfactuals} consists of
+counterfactual estimates of every available treatment.
 }
 \description{
 This algorithm trains a joint forest model to estimate the optimal treatment assignment
