+++ r-package/rjaf/NAMESPACE
@@ -14,4 +14,5 @@ importFrom(rlang,":=")
 importFrom(stats,as.formula)
 importFrom(stats,kmeans)
 importFrom(stats,predict)
+importFrom(stats,setNames)
 useDynLib(rjaf, .registration=TRUE)
