### 2025-02-13 5c9a01c
update README to reflect recent changes
