### 2025-02-18 92cd654
reduce word count per JOSS editor
