### 2025-02-24 #7750 @vissarion
Hi @wenjie2wang, you have been suggested by the authors as suitable reviewers for this submission.

In JOSS, we carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html

Would any of you be willing to review this submission for JOSS?
