### 2025-02-25 #7843 @editorialbot
[REVIEW]: rjaf: Regularized Joint Assignment Forest with Treatment Arm Clustering
**Submitting author:** @wustat (<a href="http://orcid.org/0000-0002-7642-9773">Wenbo Wu</a>)
**Repository:** https://github.com/wustat/rjaf
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.3
