### 2025-02-26 0adc6b1
change res to fitted and add xwalk to the ouput
