+++ r-package/rjaf/DESCRIPTION
@@ -29,7 +29,7 @@ LinkingTo:
     Rcpp,
     RcppArmadillo
 Depends: 
-    R (>= 3.0)
+    R (>= 3.5.0)
 License: GPL-3
 Encoding: UTF-8
 LazyData: true

[…]
