### 2025-03-02 wustat/rjaf#6 @wenjie2wang
JOSS Review Comments
This issue contains specific review comments I have for https://github.com/openjournals/joss-reviews/issues/7843.

> **Community guidelines:** Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

The package repository is lacking guideline for third parties wishing to "Contribute to the software".  Could you include contribution guidelines in the repository (e.g., inside the README) for reference?

> **State of the field:** Do the authors describe how this software compares to other commonly-used packages?

Could you briefly mention how the package {rjaf} compares to existing R packages such as {grf} in the article?

> **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?

The submitted manuscript is well-written overall and provides a clear workflow to help readers understand the underlying process, which is great. However, internal functions like `rjaf_cpp()` are mentioned multiple times, even though they are not meant to be called directly by users. To keep the workflow introduction more general and concise, could your consider avoiding references to internal function names unless necessary? Since these functions may change without affecting users, limiting their mention would improve clarity and maintainability.
