### 2025-03-06 086b70b
Create CONTRIBUTING.md
