+++ CONTRIBUTING.md
@@ -1,0 +1,43 @@
+# Contributing to the Software
+
+We welcome contributions from all users and developers! Whether you're reporting a bug, suggesting an enhancement, or submitting a pull request, your input helps improve ``rjaf``. Below are guidelines to ensure a smooth and effective contribution process.
+
+---
+
+## Reporting an Issue
+
+If you encounter a bug or have a feature request, please open an issue on our [GitHub repository](https://github.com/wustat/rjaf). To help us address your concern efficiently, follow these steps:
+
+1. **Provide a Clear Description**: Explain the issue or feature request in detail.
+2. **Include a Toy Example**: If possible, provide a minimal, reproducible example that demonstrates the problem or desired functionality.
+3. **Step-by-Step Instructions**: If a toy example isn't feasible, include detailed steps to reproduce the issue.
+4. **Environment Details**: Specify the version of ``rjaf``, your operating system, and any other relevant software dependencies.
+
+---
+
+## Submitting a Pull Request
+
+We encourage contributions through pull requests! To ensure your changes align with the goals and standards of ``rjaf``, follow these steps:
+
+1. **Create an Issue First**: Before writing code, open an issue to discuss your proposed changes. This allows for feedback and alignment with the project's direction.
+2. **Reference the Issue**: When submitting your pull request, reference the related issue to provide context.
+3. **Follow Coding Standards**: Ensure your code adheres to the project's style and conventions.
+4. **Write Clear Commit Messages**: Use descriptive commit messages to explain the purpose of each change.
+5. **Test Your Changes**: Verify that your modifications work as intended and do not introduce new issues.
+
+Once submitted, we will review your pull request and provide feedback or merge it if everything looks good.
+
+---
+
+## Contact for Support
+
+For questions, feedback, or support, feel free to reach out to:
+
+- **Xinyi Zhang**: zhang.xinyi@nyu.edu
+- **Wenbo Wu**: wenbowu@jhu.edu
+
+Alternatively, you can post an issue on our [GitHub repository](https://github.com/wustat/rjaf) for public discussion.
+
+---
+
+We appreciate your contributions and look forward to collaborating with you to make ``rjaf`` even better!
