### 2025-03-11 fc706d3
Update paper.md
GRF language added - Wenbo to wordsmith
