### 2025-03-12 c6ef9da
remove mentions of internal functions and add mention of grf
