+++ manuscript/JOSS/paper.md
@@ -9,7 +9,7 @@ tags:
 - optimal assignment
 - R
 - C++
-date: "February 18, 2025"
+date: "March 12, 2025"
 output:
   pdf_document: default
   html_document:
@@ -48,12 +48,16 @@ There is an ever-growing literature at the intersection of machine learning and 
 
 # Workflow
 
-\autoref{fig:pkg} outlines the workflow of the `rjaf` package. The process begins with partitioning the input data---consisting of outcomes, treatment arms, covariates, individual identifiers, and optional probabilities of treatment assignment---into two parts, one for model training and estimation, and the other is the held-out set on which personalized assignment rules are obtained. The `rjaf` function first checks whether outcome residualization for reducing baseline variation should be performed via the `residualize` function, using the `resid` argument. If `resid` is set to `TRUE` (the default), a new column of residualized outcomes is added to the input data and used for tree growing on the training set. Next, the `rjaf` function evaluates whether treatment clustering should be performed on the training-estimation set during tree growing using the `clus.tree.growing` argument. If `clus.tree.growing` is `TRUE`, the `rjaf_cpp` function is employed to estimate cross-validated counterfactual outcomes for the $K+1$ treatment arms, after which k-means clustering is used to learn $M+1$ treatment arm clusters. The optimal number of treatment clusters is determined using the elbow method. After clustering, the `rjaf_cpp` function is reapplied to the preprocessed data, with assignment forest fitted on $M+1$ treatment clusters and counterfactual outcomes estimated for the original $K+1$ arms. If `clus.tree.growing` is `FALSE`, the `rjaf_cpp` function is employed to estimate counterfactual outcomes for the $K+1$ arms. Lastly, `rjaf_cpp` function is used to obtain optimal treatment arms and predicted counterfactual outcomes under all treatment arms for individuals in the held-out set.
+\autoref{fig:pkg} outlines the workflow of the `rjaf` package. The process begins with partitioning the input data---consisting of outcomes, treatment arms, covariates, individual identifiers, and optional probabilities of treatment assignment---into two parts, one for model training and estimation, and the other is the held-out set on which personalized assignment rules are obtained. The `rjaf` function first checks whether outcome residualization for reducing baseline variation should be performed via the `residualize` function, using the `resid` argument. If `resid` is set to `TRUE` (the default), a new column of residualized outcomes is added to the input data and used for tree growing on the training set. Next, the `rjaf` function evaluates whether treatment clustering should be performed on the training-estimation set during tree growing using the `clus.tree.growing` argument. If `clus.tree.growing` is `TRUE`, an `Rcpp` function is employed to estimate cross-validated counterfactual outcomes for the $K+1$ treatment arms, after which k-means clustering is used to learn $M+1$ treatment arm clusters. The optimal number of treatment clusters is determined using the elbow method. After clustering, the `Rcpp` function is reapplied to the preprocessed data, with assignment forest fitted on $M+1$ treatment clusters and counterfactual outcomes estimated for the original $K+1$ arms. If `clus.tree.growing` is `FALSE`, the `Rcpp` function is employed to estimate counterfactual outcomes for the $K+1$ arms. Lastly, the `Rcpp` function is used to obtain optimal treatment arms and predicted counterfactual outcomes under all treatment arms for individuals in the held-out set.
 
 ![A sketch of the `rjaf` package. \label{fig:pkg}](pkg_sketch.pdf){height=95%}
 
[…]
