### 2025-03-12 #7843 @wustat
Thank you @wenjie2wang! My response below to your comments [wustat/rjaf#6](https://github.com/wustat/rjaf/issues/6).

> > **Community guidelines:** Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support
>
> The package repository is lacking guideline for third parties wishing to "Contribute to the software". Could you include contribution guidelines in the repository (e.g., inside the README) for reference?

A new file [CONTRIBUTING.md](https://github.com/wustat/rjaf/blob/main/CONTRIBUTING.md) was added.

> > **State of the field:** Do the authors describe how this software compares to other commonly-used packages?
>
> Could you briefly mention how the package {rjaf} compares to existing R packages such as {grf} in the article?

We noted this in the Statement of Need ("implemented through the grf package").

> > **Quality of writing:** Is the paper well written (i.e., it does not require editing for structure, language, or writing quality)?
>
> The submitted manuscript is well-written overall and provides a clear workflow to help readers understand the underlying process, which is great. However, internal functions like `rjaf_cpp()` are mentioned multiple times, even though they are not meant to be called directly by users. To keep the workflow introduction more general and concise, could your consider avoiding references to internal function names unless necessary? Since these functions may change without affecting users, limiting their mention would improve clarity and maintainability.

All mentions of internal functions were removed in the revised manuscript.
