### 2025-03-19 wustat/rjaf#7 @limengbinggz
JOSS Review: comments
These comments are for [openjournals/joss-reviews#7843](https://github.com/openjournals/joss-reviews/issues/7843).

1. In the Example, the line `head(forest.reg$fitted)` and `forest.reg$xwalk` return NULL for me, since `forest.reg` only contains two objects "res" and "counterfactuals". Please correct me if there is an updated version of the package that I should install.

2. For reproducibility of the Example, maybe consider specifying a random seed?

Paper:
1. Could you briefly describe how the method implemented in your package compares to outcome weighted learning (and related literature such as residual weighted learning, etc), and the `[policytree](https://grf-labs.github.io/policytree/)` package?

Zhao, Y., Zeng, D., Rush, A. J., & Kosorok, M. R. (2012). Estimating individualized treatment rules using outcome weighted learning. Journal of the American Statistical Association, 107(499), 1106-1118.

Zhou, X., Mayer-Hamblett, N., Khan, U., & Kosorok, M. R. (2017). "Residual weighted learning for estimating individualized treatment rules." Journal of the American Statistical Association, 112(517), 169-187.

2. Based on the last sentence of Summary and Statement of need, it seems to me that the implemented algorithm is designed for situations where heterogeneous treatment effect is present with a large number of treatment arms. I am curious about what are real examples of many treatment arms, and what distinguishes between "many" and "not many"? The Summary suggests a COVID vaccine example - but I believe only there are only a limited number of treatments.

README:
1. Could you explain the difference between Y.cf in `forest.reg$fitted` and Y_0.rjaf - Y_4.rjaf in forest.reg$counterfactuals? I am a little unclear about this sentence in the documentation - "As counterfactual outcomes present, they are also included in fitted as Y.cf. counterfactuals contains estimated counterfactual outcomes from every treatment arm."
