+++ manuscript/JOSS/paper.bib
@@ -258,3 +258,39 @@
   doi={10.32614/CRAN.package.tibble},
   url = {https://tibble.tidyverse.org}
 }
+
+@article{sverdrup2020policytree,
+  title={{policytree: Policy learning via doubly robust empirical welfare maximization over trees}},
+  author={Sverdrup, Erik and Kanodia, Ayush and Zhou, Zhengyuan and Athey, Susan and Wager, Stefan},
+  journal={Journal of Open Source Software},
+  volume={5},
+  number={50},
+  pages={2232},
+  year={2020},
+  doi={10.21105/joss.02232},
+  url={https://joss.theoj.org/papers/10.21105/joss.02232}
+}
+
+@article{zhao2012estimating,
+  title={Estimating individualized treatment rules using outcome weighted learning},
+  author={Zhao, Yingqi and Zeng, Donglin and Rush, A John and Kosorok, Michael R},
+  journal={Journal of the American Statistical Association},
+  volume={107},
+  number={499},
+  pages={1106--1118},
+  year={2012},
+  doi={10.1080/01621459.2012.695674},
+  url={https://www.tandfonline.com/doi/abs/10.1080/01621459.2012.695674}
+}
+
+@article{zhou2017residual,
+  title={Residual weighted learning for estimating individualized treatment rules},
+  author={Zhou, Xin and Mayer-Hamblett, Nicole and Khan, Umer and Kosorok, Michael R},
+  journal={Journal of the American Statistical Association},
+  volume={112},
+  number={517},
+  pages={169--187},
+  year={2017},
+  doi={10.1080/01621459.2015.1093947},
+  url={https://www.tandfonline.com/doi/abs/10.1080/01621459.2015.1093947}
+}
