+++ README.Rmd
@@ -67,9 +67,9 @@ library(rjaf)
 ```
 
 
-We use a data set simulated by `sim.data()` under the example section of `rjaf.R`. This dataset contains a total of 100 items and 5 treatment arms, with a total of 12 covariates as documented in `data.R`. After preparing the `Example_data` into training, estimation, and heldout, we can obtain regularized averages by 5 treatment arms and acquire the optimal assignment. 
+We use a dataset simulated by `sim.data()` under the example section of `rjaf.R`. This dataset contains a total of 100 rows and 5 treatment arms, with a total of 12 covariates as documented in `data.R`. After dividing the `Example_data` into training-estimation and held-out sets, we can obtain regularized averages by 5 treatment arms and optimal treatment assignments. 
 
-Our algorithm returns a list named ``forest.reg``, which includes two tibbles named ``fitted`` and ``counterfactuals``. ``fitted`` contains individual IDs, optimal treatment arms identified (`trt.rjaf`), predicted optimal outcomes (`Y.rjaf`), and treatment arm clusters (`clus.rjaf`). As counterfactual outcomes present, they are also included in ``fitted`` as `Y.cf`. ``counterfactuals`` contains estimated counterfactual outcomes from every treatment arm. If performing clustering, ``xwalk`` is included, which contains cluster number of treatment assigned by k-means algorithm.
+Our algorithm returns a list named `forest.reg`, which includes two tibbles named `fitted` and `counterfactuals`. The `fitted` contains individual IDs from the held-out set, optimal treatment arms identified (`trt.rjaf`), predicted optimal outcomes (`Y.rjaf`), and treatment arm clusters (`clus.rjaf`). In this example, since we know the counterfactual outcomes, we include those under optimal treatment arms `trt.rjaf` identified by the algorithm in `fitted` as `Y.cf`. The tibble `counterfactuals` contains estimated counterfactual outcomes for every treatment arm. If performing clustering, the tibble `xwalk` is also returned by the algorithm. `xwalk` has the treatments and their assigned cluster memberships (based on the k-means algorithm).
 
 ```{r, message = FALSE}
 library(magrittr)
@@ -77,13 +77,13 @@ library(dplyr)
 
 # prepare training, estimation, and heldout data
 data("Example_data")
-
+set.seed(1)
 # training and estimation
 data.trainest <- Example_data %>% 
-                  slice_sample (n = floor(0.5 * nrow(Example_data)))
+                  slice_sample(n=floor(0.5*nrow(Example_data)))
 # heldout
 data.heldout <- Example_data %>% 
-                  filter (!id %in% data.trainest$id)
+                  filter(!id %in% data.trainest$id)
 
 # specify variables needed
 id <- "id"; y <- "Y"; trt <- "trt";  
@@ -91,8 +91,8 @@ vars <- paste0("X", 1:3); prob <- "prob";
 
 # calling the ``rjaf`` function and implement clustering scheme
 forest.reg <- rjaf(data.trainest, data.heldout, y, id, trt, vars, 
-                   prob, clus.max = 3, 
-                   clus.tree.growing = TRUE, setseed = TRUE)
+                   prob, clus.max=3, 
+                   clus.tree.growing=TRUE, setseed=TRUE)
 
 ```
 
