+++ README.md
@@ -66,7 +66,7 @@ original arms are combined by groups) to obtain an ensemble of trees.
 The following scripts demonstrate the function `rjaf()`, which
 constructs a joint forest model to estimate the optimal treatment
 assignment by pooling information across treatment arms using a
-clustering scheme. By inputting training, estimation, and heldout data,
+clustering scheme. By inputting training, estimation, and held-out data,
 we can obtain final regularized predictions and assignments in
 `forest.reg`, where the algorithm estimates regularized averages
 separately by the original treatment arms $k \in \{0,\ldots,K\}$ and
@@ -78,22 +78,24 @@ obtains the corresponding assignment.
 library(rjaf)
 ```
 
-We use a data set simulated by `sim.data()` under the example section of
-`rjaf.R`. This dataset contains a total of 100 items and 5 treatment
+We use a dataset simulated by `sim.data()` under the example section of
+`rjaf.R`. This dataset contains a total of 100 rows and 5 treatment
 arms, with a total of 12 covariates as documented in `data.R`. After
-preparing the `Example_data` into training, estimation, and heldout, we
-can obtain regularized averages by 5 treatment arms and acquire the
-optimal assignment.
+dividing the `Example_data` into training-estimation and held-out sets,
+we can obtain regularized averages by 5 treatment arms and optimal
+treatment assignments.
 
 Our algorithm returns a list named `forest.reg`, which includes two
-tibbles named `fitted` and `counterfactuals`. `fitted` contains
-individual IDs, optimal treatment arms identified (`trt.rjaf`),
-predicted optimal outcomes (`Y.rjaf`), and treatment arm clusters
-(`clus.rjaf`). As counterfactual outcomes present, they are also
-included in `fitted` as `Y.cf`. `counterfactuals` contains estimated
-counterfactual outcomes from every treatment arm. If performing
-clustering, `xwalk` is included, which contains cluster number of
-treatment assigned by k-means algorithm.
+tibbles named `fitted` and `counterfactuals`. The `fitted` contains
+individual IDs from the held-out set, optimal treatment arms identified
+(`trt.rjaf`), predicted optimal outcomes (`Y.rjaf`), and treatment arm
+clusters (`clus.rjaf`). In this example, since we know the
+counterfactual outcomes, we include those under optimal treatment arms
+`trt.rjaf` identified by the algorithm in `fitted` as `Y.cf`. The tibble
+`counterfactuals` contains estimated counterfactual outcomes for every
+treatment arm. If performing clustering, the tibble `xwalk` is also
+returned by the algorithm. `xwalk` has the treatments and their assigned
+cluster memberships (based on the k-means algorithm).
 
 ``` r
 library(magrittr)
@@ -101,13 +103,13 @@ library(dplyr)
 
 # prepare training, estimation, and heldout data
 data("Example_data")
-
+set.seed(1)
 # training and estimation
 data.trainest <- Example_data %>% 
-                  slice_sample (n = floor(0.5 * nrow(Example_data)))
-# heldout
+                  slice_sample(n=floor(0.5*nrow(Example_data)))
+# held-out
 data.heldout <- Example_data %>% 
-                  filter (!id %in% data.trainest$id)
[…]
