### 2025-03-22 ab64810
Move id column in "counterfactuals" to the first column
