+++ README.md
@@ -134,14 +134,14 @@ head(forest.reg$fitted)
 #> 6 10    3        -26.7   16.7          2
 head(forest.reg$counterfactuals)
 #> # A tibble: 6 × 6
-#>   Y_0.rjaf Y_1.rjaf Y_2.rjaf Y_3.rjaf Y_4.rjaf id   
-#>      <dbl>    <dbl>    <dbl>    <dbl>    <dbl> <chr>
-#> 1     1.55     6.72    11.6      5.77     9.18 3    
-#> 2    -2.09     6.92     6.83    -4.10     8.56 4    
-#> 3    -1.72     1.19     7.38     3.09     4.46 5    
-#> 4     5.99    10.8     15.3     18.9     10.1  8    
-#> 5     9.59     3.15    16.7     20.7     13.3  9    
-#> 6     7.01     1.86    13.4     16.7      9.94 10
+#>   id    Y_0.rjaf Y_1.rjaf Y_2.rjaf Y_3.rjaf Y_4.rjaf
+#>   <chr>    <dbl>    <dbl>    <dbl>    <dbl>    <dbl>
+#> 1 3         1.55     6.72    11.6      5.77     9.18
+#> 2 4        -2.09     6.92     6.83    -4.10     8.56
+#> 3 5        -1.72     1.19     7.38     3.09     4.46
+#> 4 8         5.99    10.8     15.3     18.9     10.1 
+#> 5 9         9.59     3.15    16.7     20.7     13.3 
+#> 6 10        7.01     1.86    13.4     16.7      9.94
 head(forest.reg$xwalk)
 #>   cluster trt
 #> 1       1   0
