### 2025-03-24 #7843 @wustat
@limengbinggz Thank you for your comments [wustat/rjaf/issues/7](https://github.com/wustat/rjaf/issues/7)! See my response below.

> 1. In the Example, the line `head(forest.reg$fitted)` and `forest.reg$xwalk` return NULL for me, since `forest.reg` only contains two objects "res" and "counterfactuals". Please correct me if there is an updated version of the package that I should install.

The latest development version can be installed by running `devtools::install_github("wustat/rjaf", subdir = "r-package/rjaf")`.

> 2. For reproducibility of the Example, maybe consider specifying a random seed?>

We added `set.seed(1)` in the Example.

> Paper:
>
> 1. Could you briefly describe how the method implemented in your package compares to outcome weighted learning (and related literature such as residual weighted learning, etc), and the `[policytree](https://grf-labs.github.io/policytree/)` package?
>
> Zhao, Y., Zeng, D., Rush, A. J., & Kosorok, M. R. (2012). Estimating individualized treatment rules using outcome weighted learning. Journal of the American Statistical Association, 107(499), 1106-1118.
>
> Zhou, X., Mayer-Hamblett, N., Khan, U., & Kosorok, M. R. (2017). "Residual weighted learning for estimating individualized treatment rules." Journal of the American Statistical Association, 112(517), 169-187.

The outcome (or residual) weighted learning approach is based on a support vector machine, while the `policytree` package pursues doubly robust estimation in observational studies which relies on the `grf` package for nuisance function estimation. Different from these methods, the `rjaf` package is an implementation of a forest-type algorithm based on greedy recursive partitioning. We have noted these alternative approaches in the revised manuscript.

> 2. Based on the last sentence of Summary and Statement of need, it seems to me that the implemented algorithm is designed for situations where heterogeneous treatment effect is present with a large number of treatment arms. I am curious about what are real examples of many treatment arms, and what distinguishes between "many" and "not many"? The Summary suggests a COVID vaccine example - but I believe only there are only a limited number of treatments.

In theory, "many" could refer to hundreds of treatment arms or even more. In most practical settings, however, 10 treatment arms are generally deemed as many. Admittedly, there were not that many arms in the COVID-19 example. In the two PNAS examples (https://doi.org/10.1073/pnas.2101165118 and https://doi.org/10.1073/pnas.2115126119), there were 19 and 22 arms, respectively, which are considered many in practice.

> README:
>
> 1. Could you explain the difference between Y.cf in `forest.reg$fitted` and Y_0.rjaf - Y_4.rjaf in forest.reg$counterfactuals? I am a little unclear about this sentence in the documentation - "As counterfactual outcomes present, they are also included in fitted as Y.cf. counterfactuals contains estimated counterfactual outcomes from every treatment arm."

`Y.cf` in `forest.reg$fitted` includes counterfactual outcomes under optimal treatment arms `trt.rjaf` identified by the algorithm, while `Y_0.rjaf - Y_4.rjaf` through `Y_4.rjaf` in `forest.reg$counterfactuals` are estimated counterfactual outcomes for every treatment arm. We have revised the sentence to clarify this.
