### 2025-03-25 573d398
drop built-in seeding and update README.Rmd
