+++ r-package/rjaf/DESCRIPTION
@@ -1,6 +1,6 @@
 Package: rjaf
 Title: Regularized Joint Assignment Forest with Treatment Arm Clustering
-Version: 0.1.2
+Version: 0.1.3
 URL: https://github.com/wustat/rjaf
 BugReports: https://github.com/wustat/rjaf/issues
 Authors@R: 
