### 2025-03-25 0456f5e
add setseed back for rjaf.R only and update package version
