+++ r-package/rjaf/R/RcppExports.R
@@ -1,7 +1,7 @@
 # Generated by using Rcpp::compileAttributes() -> do not edit by hand
 # Generator token: 10BE3573-1514-4C36-9D1C-5A225CD40393
 
-rjaf_cpp <- function(y_trainest, y_trainest_resid, X_trainest, trt_trainest, prob_trainest, cluster_trainest, X_heldout, ntrts = 5L, nvars = 3L, lambda1 = 0.5, lambda2 = 0.5, ipw = TRUE, nodesize = 5L, ntree = 1000L, prop_train = 0.5, eps = 0.1, reg = TRUE, impute = TRUE, setseed = FALSE, seed = 1L) {
-    .Call(`_rjaf_rjaf_cpp`, y_trainest, y_trainest_resid, X_trainest, trt_trainest, prob_trainest, cluster_trainest, X_heldout, ntrts, nvars, lambda1, lambda2, ipw, nodesize, ntree, prop_train, eps, reg, impute, setseed, seed)
+rjaf_cpp <- function(y_trainest, y_trainest_resid, X_trainest, trt_trainest, prob_trainest, cluster_trainest, X_heldout, ntrts = 5L, nvars = 3L, lambda1 = 0.5, lambda2 = 0.5, ipw = TRUE, nodesize = 5L, ntree = 1000L, prop_train = 0.5, eps = 0.1, reg = TRUE, impute = TRUE) {
+    .Call(`_rjaf_rjaf_cpp`, y_trainest, y_trainest_resid, X_trainest, trt_trainest, prob_trainest, cluster_trainest, X_heldout, ntrts, nvars, lambda1, lambda2, ipw, nodesize, ntree, prop_train, eps, reg, impute)
 }
 
