+++ r-package/rjaf/man/rjaf.Rd
@@ -105,8 +105,7 @@ when the arm has no observation. If \code{FALSE}, the within-leaf average outcom
 is set to zero when the arm has no observation. The default value is \code{TRUE}.}
 
 \item{setseed}{a logical indicator. If \code{TRUE}, a seed is set through the
-argument \code{seed} below and passed to the function \code{rjaf_cpp}.
-The default value is \code{FALSE}.}
+argument \code{seed} below. The default value is \code{FALSE}.}
 
 \item{seed}{an integer used as a random seed if \code{setseed=TRUE}.
 The default value is 1.}
@@ -123,10 +122,11 @@ to the clusters, and treatment arms. Otherwise, \code{rjaf}  returns a list of t
 named \code{fitted} and \code{counterfactuals}. \code{fitted} consists of individual IDs (\code{id}),
 optimal treatment arms identified by the algorithm (\code{trt.rjaf}), treatment
 clusters (\code{clus.rjaf}) if \code{clus.tree.growing} is \code{TRUE}, and predicted optimal outcomes (\code{Y.rjaf}).
-If counterfactual outcomes are also present, they will be included
-in \code{res} along with the column of predicted outcomes (\code{Y.cf}). \code{counterfactuals} consists of
-counterfactual estimates of every available treatment. If \code{clus.tree.growing} is \code{TRUE},
-\code{rjaf} will also return a tibble \code{xwalk} that consists of \code{cluster} and \code{trt} based on k-means clustering.
+If counterfactual outcomes are also present, a column \code{Y.cf} will be included
+in \code{fitted} along with \code{Y.rjaf}. \code{counterfactuals} contains estimated
+counterfactual outcomes for every treatment arm. If \code{clus.tree.growing} is \code{TRUE},
+\code{rjaf} will also return a tibble \code{xwalk} that includes \code{trt} as the treatments
+and \code{cluster} as their assigned cluster memberships based on k-means clustering.
 }
 \description{
 This algorithm trains a joint forest model to estimate the optimal treatment assignment
@@ -174,6 +174,7 @@ sim.data <- function(n, K, gamma, sigma, prob=rep(1,K+1)/(K+1)) {
 }
 
 n <- 200; K <- 3; gamma <- 10; sigma <- 10
+set.seed(1)
 Example_data <- sim.data(n, K, gamma, sigma)
 Example_trainest <- Example_data \%>\% slice_sample(n = floor(0.5 * nrow(Example_data)))
 Example_heldout <- Example_data \%>\% filter(!id \%in\% Example_trainest$id)
