+++ manuscript/JOSS/paper.bib
@@ -259,18 +259,6 @@
   url = {https://tibble.tidyverse.org}
 }
 
-@article{sverdrup2020policytree,
-  title={{policytree: Policy learning via doubly robust empirical welfare maximization over trees}},
-  author={Sverdrup, Erik and Kanodia, Ayush and Zhou, Zhengyuan and Athey, Susan and Wager, Stefan},
-  journal={Journal of Open Source Software},
-  volume={5},
-  number={50},
-  pages={2232},
-  year={2020},
-  doi={10.21105/joss.02232},
-  url={https://joss.theoj.org/papers/10.21105/joss.02232}
-}
-
 @article{zhao2012estimating,
   title={Estimating individualized treatment rules using outcome weighted learning},
   author={Zhao, Yingqi and Zeng, Donglin and Rush, A John and Kosorok, Michael R},
