### 2025-04-09 e15b328
remove special format in paper title per JOSS editor
